Split sky-subtraction into separate module
Updated all imports to reflect the new module, and fixed PEP8 errors
